function_calling/pyproject.toml
[tool.poetry]
name = "function-calling"
version = "0.1.0"
description = ""
authors = ["thinh.dang <thinh.dang@trustingsocial.com>"]
readme = "README.md"
packages = [{include = "function_calling"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
scipy = "^1.11.2"
tiktoken = "^0.4.0"
termcolor = "^2.3.0"
tenacity = "^8.2.3"
openai = "^0.27.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
